move settings/chanell image logic to settings/channel/image
