+++ app/components/channel-header/template.hbs
@@ -5,7 +5,7 @@
 				{{channel-cover src=channel.image alt=channel.title}}
 			{{/link-to}}
 		{{else}}
-			{{#link-to "settings.channel"}}
+			{{#link-to "settings.channel.image"}}
 				{{channel-cover src=channel.image alt=channel.title}}
 			{{/link-to}}
 		{{/if}}
