+++ app/settings/channel/template.hbs
@@ -2,6 +2,7 @@
   <div class="Container Container--noPaddingMobile Container--contentCentered">
 		<nav class="Tabs Tabs--transparent Tabs--horizontal">
 			{{link-to "General" "settings.channel.index" class="Tabs-item"}}
+			{{link-to "Image" "settings.channel.image" class="Tabs-item"}}
 			{{link-to "Map" "settings.channel.map" class="Tabs-item"}}
 			{{link-to "Backup" "settings.channel.backup" class="Tabs-item"}}
 			{{link-to "Delete" "settings.channel.delete" class="Tabs-item"}}
